+++ src/app/_component/temp/Loading/index.tsx
@@ -1,18 +1,17 @@
 'use client';
 
 import * as React from 'react';
-import { ViewingPageWrap } from './style';
+import { LoadingPageWrap } from './style';
 import Image from 'next/image';
-import { Colors, Icons, Images } from '@globalStyles';
+import { Colors, Icons, Images } from '@/styles';
 import { SecureLocalStorage } from '@/hooks/useUtil';
 import { styled } from '@mui/material/styles';
 import LinearProgress, {
   linearProgressClasses,
 } from '@mui/material/LinearProgress';
 import { useEffect, useState } from 'react';
-import BackHeader from '@/app/_component/molecule/BackHeader';
 
-type props = { backUrl: string; userName?: string | null };
+type props = {};
 
 const BorderLinearProgress = styled(LinearProgress)(({ theme }) => ({
   height: 10,
@@ -27,15 +26,16 @@ const BorderLinearProgress = styled(LinearProgress)(({ theme }) => ({
   },
 }));
 
-const ViewingPage: React.FC<props> = ({ backUrl, userName }: props) => {
+const LoadingPage: React.FC<props> = ({}) => {
+  const username = SecureLocalStorage.getItem('userName');
   const [progress, setProgress] = useState(0);
 
   useEffect(() => {
     const timer = setInterval(() => {
       setProgress((prevProgress) =>
-        prevProgress >= 100 ? 0 : prevProgress + 1,
+        prevProgress >= 100 ? 0 : prevProgress + 5,
       );
-    }, 50); // 주기를 20밀리초로 설정하여 1단위로 증가
+    }, 390);
 
     return () => {
       clearInterval(timer);
@@ -43,18 +43,20 @@ const ViewingPage: React.FC<props> = ({ backUrl, userName }: props) => {
   }, []);
 
   return (
-    <ViewingPageWrap>
-      <BackHeader title={' '} url={backUrl} />
+    <LoadingPageWrap>
       <div className="container">
         <div className="top">
-          <div className="percent">{`${progress}%`}</div>
           <div className="title">
-            <p>{userName ? userName : '회원'}님</p>의 <br />
+            {username ? username : '회원'}님의 <br />
             접종 내역을 조회중이에요
+          </div>
+          <div className="subTitle">
+            예방접종등록사업을 시작한, <br />
+            2002년 이후의 예방접종기록을 확인할 수 있어요
           </div>
         </div>
         <div className="body">
-          <Image src={Images.vacgomViewing} alt={'백곰'} />
+          <Image src={Images.vacgomLoading} alt={'백곰'} />
         </div>
         <div className="bottom">
           <div className="progress">
@@ -62,8 +64,8 @@ const ViewingPage: React.FC<props> = ({ backUrl, userName }: props) => {
           </div>
         </div>
       </div>
-    </ViewingPageWrap>
+    </LoadingPageWrap>
   );
 };
 
-export default ViewingPage;
+export default LoadingPage;
