+++ src/app/_component/temp/Loading/style.ts
@@ -1,34 +1,19 @@
 import styled from '@emotion/styled';
 import { Colors, fontGenerator } from '@/styles';
-
-export const ViewingPageWrap = styled.main`
+export const LoadingPageWrap = styled.main`
   width: 100%;
   height: 100%;
+  position: absolute;
   top: 0;
   z-index: 1000;
-
+  background-color: ${Colors.Primary};
   & > .container {
-    position: relative;
-
     & > .top {
-      margin: 4vh 20px;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      height: 100%;
-      & > .percent {
-        ${fontGenerator('48px', '400', '140%')}
-        color: ${Colors.Primary};
-      }
+      margin: 5lh 20px 0 20px;
       & > .title {
-        ${fontGenerator('24px', '500', '32px')}
-        color: ${Colors.Black};
+        ${fontGenerator('24px', '700', '32px')}
+        color: ${Colors.White};
         margin-bottom: 14px;
-        white-space: nowrap;
-        p {
-          display: inline-block;
-          ${fontGenerator('24px', '700', '32px')}
-        }
       }
       & > .subTitle {
         ${fontGenerator('12px', '500', '18px')}
@@ -36,25 +21,22 @@ export const ViewingPageWrap = styled.main`
       }
     }
     & > .body {
-      padding: 20px;
-      position: relative;
-
       & > img {
         width: 100%;
-        height: 100%;
-        max-height: 60vh;
-        //position: absolute;
-        //bottom: 0;
+        position: absolute;
+        z-index: 1;
+        bottom: 53px;
       }
     }
     & > .bottom {
       background-color: ${Colors.White};
       width: 100%;
-      //height: 95px;
+      height: 95px;
+      position: absolute;
       bottom: 0;
       & > .progress {
         z-index: 2;
-        margin: 42px 20px;
+        margin: 52px 20px;
       }
     }
   }
