+++ src/app/signup/error/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import * as React from 'react';
-import { ViewingPageWrap } from './style';
+import { SignupErrorPageWrap } from './style';
 import Image from 'next/image';
 import { Colors, Icons, Images } from '@globalStyles';
 import { SecureLocalStorage } from '@/hooks/useUtil';
@@ -43,26 +43,10 @@ const SignupErrorPage: React.FC<props> = ({ backUrl, userName }: props) => {
   }, []);
 
   return (
-    <ViewingPageWrap>
-      <BackHeader title={' '} url={backUrl} />
-      <div className="container">
-        <div className="top">
-          <div className="percent">{`${progress}%`}</div>
-          <div className="title">
-            <p>{userName ? userName : '회원'}님</p>의 <br />
-            접종 내역을 조회중이에요
-          </div>
-        </div>
-        <div className="body">
-          <Image src={Images.vacgomViewing} alt={'백곰'} />
-        </div>
-        <div className="bottom">
-          <div className="progress">
-            <BorderLinearProgress variant="determinate" value={progress} />
-          </div>
-        </div>
-      </div>
-    </ViewingPageWrap>
+    <SignupErrorPageWrap>
+      <BackHeader title={' '} url={backUrl} color={'white'} />
+      <div className="container"></div>
+    </SignupErrorPageWrap>
   );
 };
