+++ src/app/signup/moreidentity/page.tsx
@@ -26,7 +26,7 @@ import { postRegister } from '@/app/_lib/postRegister';
 import WarningToast from '@/app/_component/atom/WarningToast';
 import { postLogin } from '@/app/_lib/postLogin';
 import WarningToastWrap from '@/app/_component/molecule/WorningToastWrap';
-import LoadingPage from '@/app/_component/temp/Loading';
+import LoadingPage from '@/app/_component/temp/Viewing';
 import { PATH } from '@/routes/path';
 
 export default function MoreIdentity(): React.JSX.Element {
